pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot==20.8",
    "requests>=2.32.4",
]
